更新javadoc

+++ java/00-exp-libs/exp-libs/trunk/src/test/java/exp/libs/utils/test/TestNumUtils.java
@@ -91,16 +91,16 @@ public class TestNumUtils {
 	@Test
 	public void testCompress() {
 		System.out.println(NumUtils.compress(null));
-		System.out.println(NumUtils.compress(new long[] {}));
-		System.out.println(NumUtils.compress(new long[] { 56L }));
-		System.out.println(NumUtils.compress(new long[] { 1L, 3L, 5L, 7L, 9L }));
-		System.out.println(NumUtils.compress(new long[] { 0L, 2L, 4L, 6L, 8L }));
-		System.out.println(NumUtils.compress(new long[] { 0L, 1L, 2L, 3L, 4L, 5L, 6L, 7L, 8L, 9L }));
-		System.out.println(NumUtils.compress(new long[] { 1L, 2L, 3L, 5L, 6L, 8L }));
-		System.out.println(NumUtils.compress(new long[] { 1L, 2L, 3L, 3L, 5L, 6L, 8L }));
-		System.out.println(NumUtils.compress(new long[] { 1L, 2L, 3L, 5L, 6L, 9L }));
-		System.out.println(NumUtils.compress(new long[] { 1L, 2L, 3L, 5L, 6L, 8L, 10L }));
-		System.out.println(NumUtils.compress(new long[] { 1L, 2L, 3L, 5L, 6L, 8L, 9L }));
+		System.out.println(NumUtils.compress(new int[] {}));
+		System.out.println(NumUtils.compress(new int[] { 56 }));
+		System.out.println(NumUtils.compress(new int[] { 1, 3, 5, 7, 9 }));
+		System.out.println(NumUtils.compress(new int[] { 0, 2, 4, 6, 8 }));
+		System.out.println(NumUtils.compress(new int[] { 0, 1, 2, 3, 4, 5, 6, 7, 8, 9 }));
+		System.out.println(NumUtils.compress(new int[] { 1, 2, 3, 5, 6, 8 }));
+		System.out.println(NumUtils.compress(new int[] { 1, 2, 3, 3, 5, 6, 8 }));
+		System.out.println(NumUtils.compress(new int[] { 1, 2, 3, 5, 6, 9 }));
+		System.out.println(NumUtils.compress(new int[] { 1, 2, 3, 5, 6, 8, 10 }));
+		System.out.println(NumUtils.compress(new int[] { 1, 2, 3, 5, 6, 8, 9 }));
 	}
 
 }
